move types to shared folder
