+++ backend/src/logic/Game.ts
@@ -1,4 +1,4 @@
-import { HexCoordinates } from "../types/index";
+import { HexCoordinates } from "../../../shared/types/index";
 import { Board } from "./Board";
 
 export class Game {
