+++ backend/tsconfig.json
@@ -9,5 +9,5 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["src/**/*", "tests/**/*", "index.ts"]
+  "include": ["src/**/*", "tests/**/*", "index.ts", "../shared/types"]
 }
