+++ shared/types/IBoard.ts
