+++ shared/types/hex.d.ts
