+++ shared/types/vectors.ts
